dfs-system-2/requirements.txt
# Core Python dependencies
pandas>=2.0.0
pyarrow>=12.0.0
pydantic>=2.0.0
requests>=2.31.0
tenacity>=8.2.0
python-dotenv>=1.0.0

# Async web dependencies
aiohttp>=3.8.0
aiohttp-cors>=0.7.0

# ML/AI dependencies
numpy>=1.24.0
scipy>=1.10.0
xgboost>=1.7.0
scikit-learn>=1.3.0

# Optimization dependencies
# ortools>=9.7.0  # Install separately: pip install ortools
pulp>=2.7.0

# Web UI dependencies
fastapi>=0.100.0
jinja2>=3.1.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utility dependencies
click>=8.1.0
rich>=13.0.0
aiofiles>=23.0.0
httpx>=0.24.0
